marketing/requirements.txt
anthropic>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
